pytest/requirements.txt
pytest==8.0.2
allure-pytest
pymysql
openpyxl
paramiko